+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 ## 3.2.7.1
 
 -   Fix Edit Mods button not showing on instances
+-   Fix not being able to install older packs without a logging client